+++ src/headless/log.js
@@ -41,7 +41,7 @@ const log = {
      * When using the 'error' or 'warn' loglevels, a full stacktrace will be
      * logged as well.
      * @method log#log
-     * @param { string } message - The message to be logged
+     * @param { string | Error } message - The message to be logged
      * @param { integer } level - The loglevel which allows for filtering of log messages
      */
     log (message, level, style='') {